backend/requirements.txt
flask
flask-cors
ezdxf
numpy
requests
google-genai
